# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STWV Multi-Company Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List all offers
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferBody"
      responses:
        "201":
          description: Created offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get an offer
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update an offer
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfferBody"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete an offer
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /offers/{id}/pdf:
    get:
      operationId: getOfferPdf
      tags: [offers]
      summary: Download offer as PDF
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /offers/{id}/send:
    post:
      operationId: sendOffer
      tags: [offers]
      summary: Send offer to client via email with PDF attached
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendDocumentResponse"
        "400":
          description: No client email configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Email delivery failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /offers/{id}/convert-to-contract:
    post:
      operationId: convertOfferToContract
      tags: [offers]
      summary: Convert offer to contract
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertOfferBody"
      responses:
        "201":
          description: Contract created from offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"

  # ───── CONTRACTS ─────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SendDocumentResponse:
      type: object
      properties:
        success:
          type: boolean
        email:
          type: string
      required:
        - success
        - email
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Company ───
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        legalForm:
          type: string
        country:
          type: string
        taxRegime:
          type: string
          description: "One of: vat, gst, none"
        taxNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        bankDetails:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        currency:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - legalForm
        - country
        - taxRegime
        - currency
        - isActive
        - createdAt
        - updatedAt
...
    UserBasic:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
      required:
        - id
        - email
        - role
...
    OfferLineItem:
      type: object
      properties:
        id:
          type: integer
        offerId:
          type: integer
        description:
          type: string
        quantity:
          type: string
        unitPrice:
          type: string
        amount:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - offerId
        - description
        - quantity
        - unitPrice
        - amount
        - sortOrder
        - createdAt
    Offer:
      type: object
      properties:
        id:
          type: integer
        offerNumber:
          type: string
        companyId:
          type: integer
        projectId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        title:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        status:
          type: string
          description: "One of: draft, sent, accepted, rejected, expired"
        subtotal:
          type: string
        taxAmount:
          type: string
        totalAmount:
          type: string
        currency:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/OfferLineItem"
        company:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Company"
        client:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/UserBasic"
        project:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ProjectBasic"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - offerNumber
        - companyId
        - title
        - status
        - subtotal
        - taxAmount
        - totalAmount
        - currency
        - lineItems
        - createdAt
        - updatedAt
    CreateOfferLineItemBody:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: string
        unitPrice:
          type: string
        sortOrder:
          type: integer
      required:
        - description
    CreateOfferBody:
      type: object
      properties:
        companyId:
          type: integer
        projectId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        title:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        currency:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateOfferLineItemBody"
      required:
        - companyId
        - title
    UpdateOfferBody:
      type: object
      properties:
        title:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        status:
          type: string
        currency:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateOfferLineItemBody"
    ConvertOfferBody:
      type: object
      properties:
        type:
          type: string
          description: "One of: client_service, freelancer_service"
    ProjectBasic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name

    # ─── Contracts ───
    Contract:
      type: object
      properties:
        id:
          type: integer
        contractNumber:
          type: string
        type:
          type: string
          description: "One of: client_service, freelancer_service"
        companyId:
          type: integer
        projectId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        offerId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
        status:
          type: string
          description: "One of: draft, sent, signed, cancelled"
        signedAt:
          type: ["string", "null"]
          format: date-time
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        company:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Company"
        client:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/UserBasic"
        project:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ProjectBasic"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - contractNumber
        - type
        - companyId
        - title
        - content
        - status
        - createdAt
        - updatedAt
...